WIP: drop selinux re-labelling stuff
