+++ cmd/virt-chroot/main.go
@@ -177,13 +177,6 @@ func main() {
 		NewGetEnforceCommand(), RelabelCommand(),
 	)
 
-	createTapCmd := NewCreateTapCommand()
-	createTapCmd.Flags().String("tap-name", "tap0", "the name of the tap device")
-	createTapCmd.Flags().Uint("uid", 0, "the owner of the tap device")
-	createTapCmd.Flags().Uint("gid", 0, "the group of the owner of the tap device")
-	createTapCmd.Flags().Uint32("queue-number", 0, "the number of queues to use on multi-queued devices")
-	createTapCmd.Flags().Uint32("mtu", 1500, "the link MTU of the tap device")
-
 	createMDEVCmd := NewCreateMDEVCommand()
 	createMDEVCmd.Flags().String("type", "", "the type of a mediated device")
 	createMDEVCmd.Flags().String("parent", "", "id of a parent (e.g. PCI_ID) for the new mediated device")
@@ -240,7 +233,6 @@ func main() {
 		mntCmd,
 		umntCmd,
 		selinuxCmd,
-		createTapCmd,
 		createMDEVCmd,
 		removeMDEVCmd,
 		cgroupsCmd,
