+++ pkg/network/driver/BUILD.bazel
@@ -15,7 +15,6 @@ go_library(
         "//pkg/network/dns:go_default_library",
         "//pkg/network/link:go_default_library",
         "//pkg/util/sysctl:go_default_library",
-        "//pkg/virt-handler/selinux:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
@@ -23,6 +22,7 @@ go_library(
         "//vendor/github.com/golang/mock/gomock:go_default_library",
         "//vendor/github.com/subgraph/libmacouflage:go_default_library",
         "//vendor/github.com/vishvananda/netlink:go_default_library",
+        "//vendor/golang.org/x/sys/unix:go_default_library",
         "//vendor/k8s.io/utils/net:go_default_library",
     ],
 )
