+++ pkg/virt-handler/selinux/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "context_executor.go",
         "executor.go",
         "generated_mock_executor.go",
         "labels.go",
@@ -21,14 +20,12 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "context_executor_test.go",
         "labels_test.go",
         "selinux_suite_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
-        "//vendor/github.com/golang/mock/gomock:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
     ],
